MIT-AI/MIT-MAS665-GlowGo/glowgo-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.5.0
pydantic-settings>=2.10.1
python-dotenv==1.0.0
httpx>=0.25.0
python-multipart>=0.0.6
google-auth-oauthlib==1.1.0
PyJWT>=2.9.0
google-generativeai>=0.8.0
sendgrid==6.11.0
email-validator==2.1.0
crewai>=0.70.0
crewai-tools>=0.12.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
openai>=1.51.0
litellm>=1.0.0
pytz>=2024.1
